backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "santaserver"
dynamic = ["version"]
description = "Enterprise Santa macOS security agent management server"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlmodel>=0.0.14",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.7",
    "alembic>=1.13.1",
    "fastapi-azure-auth>=4.2.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
    "httpx>=0.25.2",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.10.1",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
]

[tool.hatch.version]
path = "app/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 120
target-version = ['py313']

[tool.isort]
profile = "black"
line_length = 120

[tool.flake8]
max-line-length = 120
exclude = [".venv", "__pycache__", ".git", "build", "dist"]
extend-ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
    "flake8-pyproject>=1.2.3",
    "testcontainers[postgres]>=4.12.0",
    "autoflake>=2.3.1",
]
